+++ src/py42/services/alertrules.py
@@ -1,3 +1,7 @@
+from py42.exceptions import Py42BadRequestError
+from py42.exceptions import Py42InvalidRuleError
+from py42.exceptions import Py42NotFoundError
+from py42.exceptions import Py42UserNotOnListError
 from py42.services import BaseService
 
 
@@ -41,7 +45,11 @@ class AlertRulesService(BaseService):
 
     def add_user(self, rule_id, user_id):
         tenant_id = self._user_context.get_current_tenant_id()
-        user_details = self._user_profile_service.get_by_id(user_id)
+        try:
+            self._user_profile_service.create_if_not_exists(user_id)
+            user_details = self._user_profile_service.get_by_id(user_id)
+        except (Py42BadRequestError, Py42NotFoundError) as err:
+            raise Py42UserNotOnListError(err, user_id, "user profile")
         user_aliases = user_details.data.get(u"cloudUsernames") or []
         data = {
             u"tenantId": tenant_id,
@@ -51,7 +59,10 @@ class AlertRulesService(BaseService):
             ],
         }
         uri = u"{}{}".format(self._api_prefix, u"add-users")
-        return self._connection.post(uri, json=data)
+        try:
+            return self._connection.post(uri, json=data)
+        except Py42NotFoundError as err:
+            raise Py42InvalidRuleError(err, rule_id)
 
     def remove_user(self, rule_id, user_id):
         user_ids = [user_id]