+++ CHANGELOG.md
@@ -28,6 +28,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Fixed
 
+- Bug that occurs when trying to add a user to an alert rule who had never been on any detection list before.
+
 - Bug where an empty destination list in a device's backup set broke creation of DeviceSettings objects for that device.
 
 - Bug where all 401 Unauthorized error responses were being raised as Py42MFARequired exceptions.